+++ include/cudawrappers/cu.hpp
@@ -7,6 +7,7 @@
 #include <array>
 #include <cstddef>
 #include <exception>
+#include <iomanip>
 #include <map>
 #include <memory>
 #include <stdexcept>
@@ -125,6 +126,29 @@ class Device : public Wrapper<hipDevice_t> {
     return {name.data()};
   }
 
+  std::string getUuid() const {
+    CUuuid uuid;
+    checkCudaCall(cuDeviceGetUuid(&uuid, _obj));
+
+    // Convert a CUuuid to CUDA's string representation.
+    // The CUuuid contains an array of 16 bytes, the UUID has
+    // the form 'GPU-XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXX', with every
+    // X being an alphanumeric character.
+    std::stringstream result;
+    result << "GPU";
+
+    for (int i = 0; i < 16; ++i) {
+      if (i == 0 || i == 4 || i == 6 || i == 8 || i == 10) {
+        result << "-";
+      }
+      result << std::hex << std::setfill('0') << std::setw(2)
+             << static_cast<unsigned>(
+                    static_cast<unsigned char>(uuid.bytes[i]));
+    }
+
+    return result.str();
+  }
+
   size_t totalMem() const {
     size_t size{};
     checkCudaCall(hipDeviceTotalMem(&size, _obj));
@@ -187,6 +211,14 @@ class Context : public Wrapper<hipCtx_t> {
     return config;
   }
 
+  int occupancyMaxActiveBlocksPerMultiprocessor(int blockSize,
+                                                size_t dynamicSMemSize) {
+    int numBlocks;
+    checkCudaCall(hipModuleOccupancyMaxActiveBlocksPerMultiprocessor(
+        &numBlocks, _obj, blockSize, dynamicSMemSize));
+    return numBlocks;
+  }
+
   static void setCacheConfig(hipFuncCache_t config) {
     checkCudaCall(hipCtxSetCacheConfig(config));
   }
@@ -205,10 +237,6 @@ class Context : public Wrapper<hipCtx_t> {
     hipCtx_t context{};
     checkCudaCall(hipCtxPopCurrent(&context));
     return Context(context);
-  }
-
-  void setSharedMemConfig(hipSharedMemConfig config) {
-    checkCudaCall(hipCtxSetSharedMemConfig(config));
   }
 
   static Device getDevice() {